+++ resource/schema/resource_graph.hpp
@@ -13,10 +13,6 @@
 
 #include "resource/schema/resource_data.hpp"
 #include <boost/graph/adjacency_list.hpp>
-#include <boost/graph/graphml.hpp>
-#include <boost/graph/graphviz.hpp>
-
-#include <utility>
 
 namespace Flux {
 namespace resource_model {
@@ -52,47 +48,6 @@ using f_vtx_infra_map_t = boost::property_map<resource_graph_t, pinfra_t>::type;
 using f_vtx_iterator_t = boost::graph_traits<resource_graph_t>::vertex_iterator;
 using f_edg_iterator_t = boost::graph_traits<resource_graph_t>::edge_iterator;
 using f_out_edg_iterator_t = boost::graph_traits<resource_graph_t>::out_edge_iterator;
-
-template<class name_map, class graph_entity>
-class label_writer_t {
-   public:
-    label_writer_t (name_map &in_map) : m (in_map)
-    {
-    }
-    void operator() (std::ostream &out, const graph_entity ent) const
-    {
-        out << "[label=\"" << m[ent] << "\"]";
-    }
-
-   private:
-    name_map m;
-};
-
-class edg_label_writer_t {
-   public:
-    edg_label_writer_t (f_edg_infra_map_t &idata, subsystem_t s) : m_infra (idata), m_s (s)
-    {
-    }
-    void operator() (std::ostream &out, const edg_t &e) const
-    {
-        auto s = m_infra[e].member_of[m_s];
-        if (!s) {
-            out << "[label=\"" << m_s << "\"]";
-        } else {
-            for (auto const &key : m_infra[e].member_of.key_range ()) {
-                if (m_infra[e].member_of[key]) {
-                    out << "[label=\"" << key << "\"]";
-                    return;
-                }
-            }
-            out << "[label=\"unknown\"]";
-        }
-    }
-
-   private:
-    f_edg_infra_map_t m_infra;
-    subsystem_t m_s;
-};
 
 }  // namespace resource_model
 }  // namespace Flux
