+++ resource/utilities/resource-query.cpp
@@ -25,7 +25,10 @@ extern "C" {
 #include <memory>
 #include <editline/readline.h>
 #include <boost/algorithm/string.hpp>
+#include <boost/graph/graphviz.hpp>
+#include <boost/graph/graphml.hpp>
 #include <filesystem>
+#include <readers/resource_reader_factory.hpp>
 #include "resource/utilities/command.hpp"
 #include "resource/store/resource_graph_store.hpp"
 #include "resource/policies/dfu_match_policy_factory.hpp"
@@ -184,6 +187,50 @@ OPTIONS:
 )";
     exit (code);
 }
+
+namespace Flux::resource_model {
+template<class name_map, class graph_entity>
+class label_writer_t {
+   public:
+    label_writer_t (name_map &in_map) : m (in_map)
+    {
+    }
+    void operator() (std::ostream &out, const graph_entity ent) const
+    {
+        out << "[label=\"" << m[ent] << "\"]";
+    }
+
+   private:
+    name_map m;
+};
+
+class edg_label_writer_t {
+   public:
+    edg_label_writer_t (f_edg_infra_map_t &idata, subsystem_t s) : m_infra (idata), m_s (s)
+    {
+    }
+    void operator() (std::ostream &out, const edg_t &e) const
+    {
+        auto s = m_infra[e].member_of[m_s];
+        if (!s) {
+            out << "[label=\"" << m_s << "\"]";
+        } else {
+            for (auto const &key : m_infra[e].member_of.key_range ()) {
+                if (m_infra[e].member_of[key]) {
+                    out << "[label=\"" << key << "\"]";
+                    return;
+                }
+            }
+            out << "[label=\"unknown\"]";
+        }
+    }
+
+   private:
+    f_edg_infra_map_t m_infra;
+    subsystem_t m_s;
+};
+
+}  // namespace Flux::resource_model
 
 static void set_default_params (std::shared_ptr<resource_context_t> &ctx)
 {
